// packages/animanga/lib/anime/anime-schedule/anime-schedule.types.ts
export type Schedule = {
	title: string;
	romaji: string;
	english: string;
	native: string;
	route: string;
	delayedFrom: string;
	delayedUntil: string;
	status: string;
	episodeDate: string;
	episodeNumber: number;
	episodes: number;
	lengthMin: number;
	imageVersionRoute: string;
	airingStatus: string;
};